requirements.txt
accelerate==0.23.0
appdirs==1.4.4
loralib
bitsandbytes==0.41.1
datasets==2.12.0
fire==0.5.0
peft==0.5.0
transformers==4.34.0
gradio
scipy
numpy
pandas
pyyaml
biopython
